fix: fit windows after AppKit frame restoration
